requirements.txt
Faker==10.0.0
imutils==0.5.4
matplotlib==3.5.1
numpy==1.26.2
opencv_python==4.5.5.62
scikit_learn==1.0.2
setuptools==52.0.0
texttable==1.6.4
requests~=2.23.0
pyyaml~=5.3.1
responses~=0.10.15
pytest
pytest-cov~=2.10.0
pytest-mock~=3.1.1
pytest-httpserver~=0.3.4
black~=19.10b
flake8~=3.8.3
flake8-black~=0.2.0
isort~=5.1.4
coverage~=5.1
validators
pyspark[sql]~=3.0.0
pycryptodome~=3.9.8
pyarrow==6.0.1
opencv_python==4.5.5.62
pdf2image==1.16.0
Pillow==9.0.1
pypdf==3.4.1
pytesseract==0.3.9
transformers==4.17.0
aspose-words==23.12.0
pytest-shutil==1.7.0

poppler-utils
numpy
pandas
